agent_template_56723/requirements.txt
```
python-dotenv
langchain-core
langgraph
```